support subcontext
temp

temp

deprecated

temp

support init event

support i18n

boot subContext when SubContextView component firstly enabled

temp
