+++ cocos2d/core/components/WXSubContextView.js
@@ -23,7 +23,7 @@
  THE SOFTWARE.
  ****************************************************************************/
 
-const Component = require('./CCComponent');
+const SubContextView = require('./SubContextView');
 
 /**
  * !#en WXSubContextView is a view component which controls open data context viewport in WeChat game platform.<br/>
@@ -48,181 +48,15 @@ const Component = require('./CCComponent');
  * 唯一需要注意的是，当子域节点的包围盒发生改变时，开发者需要使用 `updateSubContextViewport` 来手动更新子域视窗。
  * @class WXSubContextView
  * @extends Component
+ * @deprecated since v2.4
  */
 let WXSubContextView = cc.Class({
     name: 'cc.WXSubContextView',
-    extends: Component,
+    extends: SubContextView,
 
     editor: CC_EDITOR && {
-        menu: 'i18n:MAIN_MENU.component.others/WXSubContextView',
+        menu: 'i18n:MAIN_MENU.component.others/WXSubContextView (Legacy)',
         help: 'i18n:COMPONENT.help_url.wx_subcontext_view'
-    },
-
-    properties: {
-        _fps: 60,
-
-        fps: {
-            get () {
-                return this._fps;
-            },
-            set (value) {
-                if (this._fps === value) {
-                    return;
-                }
-                this._fps = value;
-                this._updateInterval = 1 / value;
-                this._updateSubContextFrameRate();
-            },
-            tooltip: CC_DEV && 'i18n:COMPONENT.wx_subcontext_view.fps'
-        }
-    },
-
-    ctor () {
-        this._sprite = null;
-        this._tex = new cc.Texture2D();
-        this._context = null;
-        this._updatedTime = performance.now();
-        this._updateInterval = 0;
-    },
-
-    onLoad () {
-        // Setup subcontext canvas size
-        if (wx.getOpenDataContext) {
-            this._updateInterval = 1000 / this._fps;
-            this._context = wx.getOpenDataContext();
-            // reset sharedCanvas width and height
-            this.reset();
-
-            this._tex.setPremultiplyAlpha(true);
-            this._tex.initWithElement(sharedCanvas);
-
-            this._sprite = this.node.getComponent(cc.Sprite);
-            if (!this._sprite) {
-                this._sprite = this.node.addComponent(cc.Sprite);
-                this._sprite.srcBlendFactor = cc.macro.BlendFactor.ONE;
-            }
-            this._sprite.spriteFrame = new cc.SpriteFrame(this._tex);
-        }
-        else {
-            this.enabled = false;
-        }
-    },
-
-    /**
-     * !#en Reset open data context size and viewport
-     * !#zh 重置开放数据域的尺寸和视窗
-     * @method reset
-     */
-    reset () {
-        if (this._context) {
-            this.updateSubContextViewport();
-            let sharedCanvas = this._context.canvas;
-            if (sharedCanvas) {
-                sharedCanvas.width = this.node.width;
-                sharedCanvas.height = this.node.height;
-            }
-        }
-    },
-
-    onEnable () {
-        this._runSubContextMainLoop();
-        this._registerNodeEvent();
-        this._updateSubContextFrameRate();
-        this.updateSubContextViewport();
-    },
-
-    onDisable () {
-        this._unregisterNodeEvent();
-        this._stopSubContextMainLoop();
-    },
-
-    update (dt) {
-        let calledUpdateMannually = (dt === undefined);
-        if (calledUpdateMannually) {
-            this._context && this._context.postMessage({
-                fromEngine: true,
-                event: 'step',
-            });
-            this._updateSubContextTexture();
-            return;
-        }
-        let now = performance.now();
-        let deltaTime = (now - this._updatedTime);
-        if (deltaTime >= this._updateInterval) {
-            this._updatedTime += this._updateInterval;
-            this._updateSubContextTexture();
-        }
-    },
-
-    _updateSubContextTexture () {
-        if (!this._tex || !this._context) {
-            return;
-        }
-        this._tex.initWithElement(this._context.canvas);
-        this._sprite._activateMaterial();
-    },
-
-    /**
-     * !#en Update the sub context viewport manually, it should be called whenever the node's bounding box changes.
-     * !#zh 更新开放数据域相对于主域的 viewport，这个函数应该在节点包围盒改变时手动调用。
-     * @method updateSubContextViewport
-     */
-    updateSubContextViewport () {
-        if (this._context) {
-            let box = this.node.getBoundingBoxToWorld();
-            let sx = cc.view._scaleX;
-            let sy = cc.view._scaleY;
-            this._context.postMessage({
-                fromEngine: true,
-                event: 'viewport',
-                x: box.x * sx + cc.view._viewportRect.x,
-                y: box.y * sy + cc.view._viewportRect.y,
-                width: box.width * sx,
-                height: box.height * sy
-            });
-        }
-    },
-
-    _registerNodeEvent () {
-        this.node.on('position-changed', this.updateSubContextViewport, this);
-        this.node.on('scale-changed', this.updateSubContextViewport, this);
-        this.node.on('size-changed', this.updateSubContextViewport, this);
-    },
-
-    _unregisterNodeEvent () {
-        this.node.off('position-changed', this.updateSubContextViewport, this);
-        this.node.off('scale-changed', this.updateSubContextViewport, this);
-        this.node.off('size-changed', this.updateSubContextViewport, this);
-    },
-
-    _runSubContextMainLoop () {
-        if (this._context) {
-            this._context.postMessage({
-                fromEngine: true,
-                event: 'mainLoop',
-                value: true,
-            });
-        }
-    },
-
-    _stopSubContextMainLoop () {
-        if (this._context) {
-            this._context.postMessage({
-                fromEngine: true,
-                event: 'mainLoop',
-                value: false,
-            });
-        }
-    },
-
-    _updateSubContextFrameRate () {
-        if (this._context) {
-            this._context.postMessage({
-                fromEngine: true,
-                event: 'frameRate',
-                value: this._fps,
-            });
-        }
     },
 });
